requirements.txt
certifi==2021.5.30
chardet==4.0.0
dnspython==2.1.0
idna==2.10
mysql-connector==2.2.9
numpy==1.21.0
pandas==1.2.5
psycopg2-binary==2.9.1
pymongo==3.11.4
python-dateutil==2.8.2
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
six==1.16.0
urllib3==1.26.6
